Completed final Submission (Reupload of Previous Commit submitted at 11:58)
I did not coordinate with my group and ended up uploading FinalLab (1).ipynb after James uploaded the a completed version. Because of that I have spent the last 30 minutes attempting to undo my file upload (I do not think that is possible). Because I am unable to , I am taking James' completed file and uploading it. He uploaded this before midnight so please do not dock any of the other team members for being late. If that does need to happen I am totally willing to take the dock in points since I created this mess of back and forth files.